+++ app/src/main/java/com/example/returnpals/Application.kt
@@ -1,7 +1,7 @@
 package com.example.returnpals
 
 import android.app.Application
-import com.example.returnpals.services.Backend
+import com.example.returnpals.viewmodel.Backend
 
 class ReturnPals : Application() {
     override fun onCreate() {
