+++ app/src/main/java/com/example/returnpals/composetools/pickup/PackagesUI.kt
@@ -57,7 +57,7 @@ import com.example.returnpals.PackageInfo
 import com.example.returnpals.composetools.IconManager
 import com.example.returnpals.composetools.ScheduleReturnScaffold
 import com.example.returnpals.composetools.getBackGroundColor
-import com.example.returnpals.services.Backend
+import com.example.returnpals.viewmodel.Backend
 import com.example.returnpals.toNiceString
 import java.io.File
